backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-backend"
version = "0.1.0"
description = "Multi-tenant RAG knowledge management platform backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    "llama-index-core>=0.12.0",
    "qdrant-client>=1.12.0",
    "markitdown[all]>=0.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "aiofiles>=24.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Local sparse embedding via FastEmbed (only needed when SPARSE_EMBEDDING_MODE=local)
local-sparse = [
    "fastembed>=0.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
